+++ layers/CMSSchema/packages/custompost-tag-mutations/src/TypeAPIs/CustomPostTagTypeMutationAPIInterface.php
@@ -7,7 +7,7 @@ namespace PoPCMSSchema\CustomPostTagMutations\TypeAPIs;
 interface CustomPostTagTypeMutationAPIInterface
 {
     /**
-     * @param string[] $tags
+     * @param array<string|int> $tags List of tags by ID, slug, or a combination of them
      */
     public function setTags(int|string $postID, array $tags, bool $append = false): void;
 }