+++ layers/GatoGraphQLForWP/plugins/gato-graphql/CHANGELOG.md
@@ -19,6 +19,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Added support for the "Multi-Field Directives" custom feature
 - Mutations `createPost`, `updatePost`, `addCommentToCustomPost` (and others) now receive a oneof input object for content
 - Mutations `createPost` and `updatePost` now have input `authorID`, as a “sensitive” data element
+- Mutations `setTagsOnPost`, `createPost` and `updatePost` now receive a oneof input object for tags
 - The Settings page has been re-designed, featuring 2-level organization and tabs displayed vertically
 - Reset settings, and choose to use restrictive or non-restrictive default settings
 - Configuration blocks in the the Schema Configuration CPT editor can be removed (and added again)
@@ -44,6 +45,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Breaking changes
 
 - Must update mutations `createPost`, `updatePost`, `addCommentToCustomPost` (and others)
+- Must update mutations `setTagsOnPost`, `createPost` and `updatePost`
 - Non-restrictive Settings values are used by default
 - Env var `ENABLE_UNSAFE_DEFAULTS` has been removed and `SETTINGS_OPTION_ENABLE_RESTRICTIVE_DEFAULT_BEHAVIOR` added in its place, to indicate to use the restrictive Settings values by default
 - Renamed plugin to "Gato GraphQL"